Fix tests

+++ crates/cubecl/benches/matmul.rs
@@ -142,7 +142,7 @@ fn main() {
         //     Default::default(),
         //     matmul::Strategy::CmmaOld(PredefinedCmmaConfig::M128K16.into()),
         // );
-        run::<cubecl::cuda::CudaRuntime, half::f16>(
+        run::<cubecl::cuda::CudaRuntime, f32>(
             Default::default(),
             matmul::Strategy::Accelerated,
         );
